fix(bazel): Escape single quotes in genrule cmd
Arguments for custom target commands, which are converted to Bazel
genrules, are joined into a single string. If any of these
arguments contain a single quote ('), it can break the shell
parsing of the `cmd` attribute in the genrule, leading to a
build failure.

This change escapes single quotes by replacing them with \',
ensuring they are treated as literal characters by the shell.

Bug: 428923963
Change-Id: Ie64732e8f3b24ef5bebbbf422999eaa2b84c5921