Fix amc to work with Qemu
Remove the recently duplicated (but flawed) meson templating logic, which was being applied before the original and so making it redundant.
Move the original logic earlier in the process and retain the @DEPFILE@
substitution which wasn't covered originally.

We should really make sure that any changes to amc do not cause
erroneous diffs to the projects that use it - probably by adding more
testing.

Change-Id: I51d7107105d7c91abe9e0ec6c0e17d8e9593d894